refactored common.py and some other files
